widget/tests/test_bug673301.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

<script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"/>
<script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

<body  xmlns="http://www.w3.org/1999/xhtml">
<p id="display"></p>
<div id="content" style="display: none"/>
</body>

<script type="application/javascript">
var clipboard = Components.classes["@mozilla.org/widget/clipboard;1"]
                          .getService(Components.interfaces.nsIClipboard);

var transferable = Components.classes['@mozilla.org/widget/transferable;1']
                             .createInstance(Components.interfaces.nsITransferable);

transferable.addDataFlavor("text/unicode");
transferable.setTransferData("text/unicode", document, 4);

clipboard.setData(transferable, null, Components.interfaces.nsIClipboard.kGlobalClipboard);

SimpleTest.ok(true, "Didn't crash setting non-text data for text/unicode type");
</script>
</window>
